Move "Recently Added Resources" to new panel re: #1574
