+++ arches/app/media/js/views/base-manager.js
@@ -4,7 +4,8 @@ define([
     'knockout',
     'backbone',
     'views/page-view',
-    'view-data'
+    'view-data',
+    'bindings/datatable'
 ], function($, _, ko, Backbone, PageView, data) {
 
     var BaseManager = PageView.extend({
@@ -46,6 +47,20 @@ define([
               if (item) {
                 ko.applyBindingsToNode(option, {disable: item.disable_instance_creation}, item);
               }
+            };
+            options.viewModel.editResource = function(url, vm, e){
+                e.stopPropagation();
+                this.navigate(url)
+            };
+            options.viewModel.resourceTableConfig = {
+                "responsive": true,
+                "language": {
+                    "paginate": {
+                        "previous": '<i class="fa fa-angle-left"></i>',
+                        "next": '<i class="fa fa-angle-right"></i>'
+                    }
+                },
+                "order": [[ 3, "desc" ]]
             };
 
             PageView.prototype.constructor.call(this, options);
