+++ arches/app/media/js/views/resource.js
@@ -4,8 +4,7 @@ require([
     'knockout',
     'arches',
     'views/base-manager',
-    'bindings/chosen',
-    'bindings/datatable'
+    'bindings/chosen'
 ], function($, _, ko, arches, BaseManagerView) {
     /**
     * a BaseManagerView representing the resource listing and recent edits pages
@@ -15,23 +14,8 @@ require([
             var self = this;
 
             _.defaults(this.viewModel, {
-                showResources: ko.observable(true),
                 showFind: ko.observable(false),
-                graphId: ko.observable(null),
-                editResource: function(url, vm, e){
-                    e.stopPropagation();
-                    this.navigate(url)
-                },
-                tableConfig: {
-                    "responsive": true,
-                    "language": {
-                        "paginate": {
-                            "previous": '<i class="fa fa-angle-left"></i>',
-                            "next": '<i class="fa fa-angle-right"></i>'
-                        }
-                    },
-                    "order": [[ 3, "desc" ]]
-                }
+                graphId: ko.observable(null)
             });
 
             this.viewModel.graphId.subscribe(function (graphid) {
