+++ arches/app/templates/views/resource.htm
@@ -57,8 +57,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
             <!-- Resource/Graph Toggle -->
             <div class="resource-tools relative" data-bind="visible: !showFind()">
-                <a class="resource-grid-title" href="javascript:void(0);" data-bind="css: {'active': showResources()}, click: function () { showResources(true) }">{% trans "Resources" %}</a>
-                <a class="resource-grid-title" href="javascript:void(0);" data-bind="css: {'active': !showResources()}, click: function () { showResources(false) }">{% trans "Recently Added" %}</a>
+                <span class="resource-grid-title">{% trans "Resources" %}</span>
             </div>
 
         </div>
@@ -69,7 +68,7 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
     <!-- Default Page content-->
     <div id="page-content" style="padding: 0 50px 160px; margin-top: -1px;">
 
-        <div id="resource-list" class="report-image-grid" style="margin-top: 5px;" data-bind="visible: showResources()">
+        <div id="resource-list" class="report-image-grid" style="margin-top: 5px;">
 
             <!-- ko foreach: allGraphs -->
             <!-- ko if: isresource -->
@@ -104,40 +103,6 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
 
         </div>
 
-
-        <!-- Resource Instances -->
-        <div id="resource-records" class="report-image-grid" style="display: none;" data-bind="visible: !showResources()">
-
-            <!-- Resource Records Table -->
-            <div class="panel">
-                <div class="panel-body recent-additions-container">
-                    <table data-bind="datatable: tableConfig" class="table table-striped table-bordered" cellspacing="0" width="100%">
-                        <thead>
-                            <tr>
-                                <th>Resource Id</th>
-                                <th>Resource Name</th>
-                                <th>Resource Type</th>
-                                <th class="min-desktop">Created</th>
-                                <th class="min-tablet"></th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {% for resource in resource_instances %}
-                            <tr>
-                                <td><a href="" class="regular-link" data-bind="click: editResource.bind($data, '{% url 'resource_editor' resource.pk %}')">{{resource.pk}}</a></td>
-                                <td>{{resource.name}}</td>
-                                <td>{{resource.graph.name}}</td>
-                                <td>{{resource.createdtime}}</td>
-                                <td>
-                                    <a href="#" data-bind="click: function () { window.open('{% url 'resource_report' resource.pk %}'); }, clickBubble: false" class="btn btn-primary">{% trans "View Report" %}</a>
-                                </td>
-                            </tr>
-                            {% endfor %}
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
 
     </div>
 </div>
